+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - New: Application API endpoint that returns all known characters from a list of character IDs.
 - New: Application API endpoint that returns all known characters from one corporation. 
 - New: Added the date when a character was added to the database to the list of characters (UI).
+- New: Required groups are now checked when adding new members to a group.
 - New: Implemented API and UI to add, rename and delete watchlists.
 - New: Lock for the watchlist settings implemented: If enabled, only watchlist admins can add and remove alliances 
   and corporations from the list of watched alliances and corporations. Watchlist managers can still edit the